+++ src/converse-rosterview.js
@@ -263,7 +263,7 @@ converse.plugins.add('converse-rosterview', {
             className: 'roster-filter-form',
             events: {
                 "keydown .roster-filter": "liveFilter",
-                "submit form.roster-filter-form": "submitFilter",
+                "submit": "submitFilter",
                 "click .clear-input": "clearFilter",
                 "click .filter-by span": "changeTypeFilter",
                 "change .state-type": "changeChatStateFilter"
@@ -926,7 +926,6 @@ converse.plugins.add('converse-rosterview', {
             },
 
             onContactChange (contact) {
-                this.updateChatBox(contact)
                 this.update();
                 if (_.has(contact.changed, 'subscription')) {
                     if (contact.changed.subscription === 'from') {
@@ -942,18 +941,6 @@ converse.plugins.add('converse-rosterview', {
                     this.addContactToGroup(contact, HEADER_REQUESTING_CONTACTS);
                 }
                 this.updateFilter();
-            },
-
-            updateChatBox (contact) {
-                if (!this.model.chatbox) {
-                    return this;
-                }
-                const changes = {};
-                if (_.has(contact.changed, 'status')) {
-                    changes.status = contact.get('status');
-                }
-                this.model.chatbox.save(changes);
-                return this;
             },
 
             getGroup (name) {